New Update
Exporting capability to PNG, some Alpha channels may be wrong for some formats
but this app was never intended for that it was a viewer and it extracts the gtt files.
Some GTT NTF7 have only 4 images rest are black and another did have 8.
For the older clients all 8 seem to load fine.
Hero Online also seems to load fine, also the GTT.
Readded better panning and zooming.
Added a status bar text with last directory filename and number of textures.
Added reading and saving to settings.txt to read last known directory because this was really obnoxious without it.
The reason why this application works so great is because it gets the individiual NTF header offsets in a file so it doesn't matter what's between them before them or after them.
The only issue seems to be dungeon(_a).dxt maybe due to indoor encryption but I don't feel like figuring it out if anyone knows you are free to contribute to do so.
I've tried adding every possible format and supporting them all , but 100% functionality is a PITA, some are hard to find.